expose prune repo defaults function and move it to the org package
