+++ prow/cmd/peribolos/main.go
@@ -363,7 +363,7 @@ func dumpOrgConfig(client dumpClient, orgName string, ignoreSecretTeams bool) (*
 			return nil, fmt.Errorf("failed to get repo: %v", err)
 		}
 		logrus.WithField("repo", full.FullName).Debug("Recording repo.")
-		out.Repos[full.Name] = pruneRepoDefaults(org.Repo{
+		out.Repos[full.Name] = org.PruneRepoDefaults(org.Repo{
 			Description:      &full.Description,
 			HomePage:         &full.Homepage,
 			Private:          &full.Private,
@@ -379,39 +379,6 @@ func dumpOrgConfig(client dumpClient, orgName string, ignoreSecretTeams bool) (*
 	}
 
 	return &out, nil
-}
-
-// pruneRepoDefaults finds values in org.Repo config that matches the default
-// values replaces them with nil pointer. This reduces the size of an org dump
-// by omitting the fields that would be set to the same value when not set at all.
-// See https://developer.github.com/v3/repos/#edit
-func pruneRepoDefaults(repo org.Repo) org.Repo {
-	pruneString := func(p **string, def string) {
-		if *p != nil && **p == def {
-			*p = nil
-		}
-	}
-	pruneBool := func(p **bool, def bool) {
-		if *p != nil && **p == def {
-			*p = nil
-		}
-	}
-
-	pruneString(&repo.Description, "")
-	pruneString(&repo.HomePage, "")
-
-	pruneBool(&repo.Private, false)
-	pruneBool(&repo.HasIssues, true)
-	// Projects' defaults depend on org setting, do not prune
-	pruneBool(&repo.HasWiki, true)
-	pruneBool(&repo.AllowRebaseMerge, true)
-	pruneBool(&repo.AllowSquashMerge, true)
-	pruneBool(&repo.AllowMergeCommit, true)
-
-	pruneBool(&repo.Archived, false)
-	pruneString(&repo.DefaultBranch, "master")
-
-	return repo
 }
 
 type orgClient interface {
