+++ prow/cmd/peribolos/main_test.go
@@ -3083,72 +3083,3 @@ func TestNewRepoUpdateRequest(t *testing.T) {
 		})
 	}
 }
-
-func TestPruneRepoDefaults(t *testing.T) {
-	empty := ""
-	nonEmpty := "string that is not empty"
-	yes := true
-	no := false
-	master := "master"
-	notMaster := "not-master"
-	testCases := []struct {
-		description string
-		repo        org.Repo
-		expected    org.Repo
-	}{
-		{
-			description: "default values are pruned",
-			repo: org.Repo{
-				Description:      &empty,
-				HomePage:         &empty,
-				Private:          &no,
-				HasIssues:        &yes,
-				HasProjects:      &yes,
-				HasWiki:          &yes,
-				AllowSquashMerge: &yes,
-				AllowMergeCommit: &yes,
-				AllowRebaseMerge: &yes,
-				DefaultBranch:    &master,
-				Archived:         &no,
-			},
-			expected: org.Repo{HasProjects: &yes},
-		},
-		{
-			description: "non-default values are not pruned",
-			repo: org.Repo{
-				Description:      &nonEmpty,
-				HomePage:         &nonEmpty,
-				Private:          &yes,
-				HasIssues:        &no,
-				HasProjects:      &no,
-				HasWiki:          &no,
-				AllowSquashMerge: &no,
-				AllowMergeCommit: &no,
-				AllowRebaseMerge: &no,
-				DefaultBranch:    &notMaster,
-				Archived:         &yes,
-			},
-			expected: org.Repo{Description: &nonEmpty,
-				HomePage:         &nonEmpty,
-				Private:          &yes,
-				HasIssues:        &no,
-				HasProjects:      &no,
-				HasWiki:          &no,
-				AllowSquashMerge: &no,
-				AllowMergeCommit: &no,
-				AllowRebaseMerge: &no,
-				DefaultBranch:    &notMaster,
-				Archived:         &yes,
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.description, func(t *testing.T) {
-			pruned := pruneRepoDefaults(tc.repo)
-			if !reflect.DeepEqual(tc.expected, pruned) {
-				t.Errorf("%s: result differs from expected:\n", diff.ObjectReflectDiff(tc.expected, pruned))
-			}
-		})
-	}
-}
