+++ prow/config/org/BUILD.bazel
@@ -26,4 +26,5 @@ go_test(
     name = "go_default_test",
     srcs = ["org_test.go"],
     embed = [":go_default_library"],
+    deps = ["@io_k8s_apimachinery//pkg/util/diff:go_default_library"],
 )
