+++ prow/config/org/org.go
@@ -139,3 +139,36 @@ func (p *Privacy) UnmarshalText(text []byte) error {
 	*p = v
 	return nil
 }
+
+// PruneRepoDefaults finds values in org.Repo config that matches the default
+// values replaces them with nil pointer. This reduces the size of an org dump
+// by omitting the fields that would be set to the same value when not set at all.
+// See https://developer.github.com/v3/repos/#edit
+func PruneRepoDefaults(repo Repo) Repo {
+	pruneString := func(p **string, def string) {
+		if *p != nil && **p == def {
+			*p = nil
+		}
+	}
+	pruneBool := func(p **bool, def bool) {
+		if *p != nil && **p == def {
+			*p = nil
+		}
+	}
+
+	pruneString(&repo.Description, "")
+	pruneString(&repo.HomePage, "")
+
+	pruneBool(&repo.Private, false)
+	pruneBool(&repo.HasIssues, true)
+	// Projects' defaults depend on org setting, do not prune
+	pruneBool(&repo.HasWiki, true)
+	pruneBool(&repo.AllowRebaseMerge, true)
+	pruneBool(&repo.AllowSquashMerge, true)
+	pruneBool(&repo.AllowMergeCommit, true)
+
+	pruneBool(&repo.Archived, false)
+	pruneString(&repo.DefaultBranch, "master")
+
+	return repo
+}
