+++ prow/config/org/org_test.go
@@ -18,7 +18,10 @@ package org
 
 import (
 	"encoding/json"
+	"reflect"
 	"testing"
+
+	"k8s.io/apimachinery/pkg/util/diff"
 )
 
 func TestPrivacy(t *testing.T) {
@@ -57,5 +60,74 @@ func TestPrivacy(t *testing.T) {
 		case err == nil && *tc.expected != actual:
 			t.Errorf("%s: actual %v != expected %v", tc.input, tc.expected, actual)
 		}
+	}
+}
+
+func TestPruneRepoDefaults(t *testing.T) {
+	empty := ""
+	nonEmpty := "string that is not empty"
+	yes := true
+	no := false
+	master := "master"
+	notMaster := "not-master"
+	testCases := []struct {
+		description string
+		repo        Repo
+		expected    Repo
+	}{
+		{
+			description: "default values are pruned",
+			repo: Repo{
+				Description:      &empty,
+				HomePage:         &empty,
+				Private:          &no,
+				HasIssues:        &yes,
+				HasProjects:      &yes,
+				HasWiki:          &yes,
+				AllowSquashMerge: &yes,
+				AllowMergeCommit: &yes,
+				AllowRebaseMerge: &yes,
+				DefaultBranch:    &master,
+				Archived:         &no,
+			},
+			expected: Repo{HasProjects: &yes},
+		},
+		{
+			description: "non-default values are not pruned",
+			repo: Repo{
+				Description:      &nonEmpty,
+				HomePage:         &nonEmpty,
+				Private:          &yes,
+				HasIssues:        &no,
+				HasProjects:      &no,
+				HasWiki:          &no,
+				AllowSquashMerge: &no,
+				AllowMergeCommit: &no,
+				AllowRebaseMerge: &no,
+				DefaultBranch:    &notMaster,
+				Archived:         &yes,
+			},
+			expected: Repo{Description: &nonEmpty,
+				HomePage:         &nonEmpty,
+				Private:          &yes,
+				HasIssues:        &no,
+				HasProjects:      &no,
+				HasWiki:          &no,
+				AllowSquashMerge: &no,
+				AllowMergeCommit: &no,
+				AllowRebaseMerge: &no,
+				DefaultBranch:    &notMaster,
+				Archived:         &yes,
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.description, func(t *testing.T) {
+			pruned := PruneRepoDefaults(tc.repo)
+			if !reflect.DeepEqual(tc.expected, pruned) {
+				t.Errorf("%s: result differs from expected:\n", diff.ObjectReflectDiff(tc.expected, pruned))
+			}
+		})
 	}
 }
